+++ Changelog.md
@@ -66,6 +66,7 @@
 * Drop support for GHC 7.10.3.
 * The following functions in 'Streamly.Prelude' have been renamed:
     * `yield` to `value`
+    * `yieldM` to `valueM`
 * The following functions in `Streamly.Data.Fold` have been renamed:
     * `mapM` to `rmapM`
     * `sequence` to `rsequence`